Actualizacion Facturacion Eletronica

+++ l10n_ec_sri/src/lib/firm.sh
@@ -7,8 +7,13 @@ mkdir -p "$path_odoo$folder_no_firm"
 mkdir -p "$path_odoo$folder_firm"
 
 for file in $(ls "$path_odoo$folder_no_firm"); do
-   rm "$path_odoo$folder_firm/$file.xml"
-   source "$path_odoo$folder_no_firm/$file" > "$path_odoo$folder_firm/$file.xml"
+   _FILE="$path_odoo$folder_firm/$file.xml"
+   if test -f "$_FILE"; then
+      rm "$path_odoo$folder_firm/$file.xml"
+   fi
+
+   #source "$path_odoo$folder_no_firm/$file" > "$path_odoo$folder_firm/$file.xml"
+   source "$path_odoo$folder_no_firm/$file" > "$_FILE"
    rm "$path_odoo$folder_no_firm/$file"
 done
